+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Change Log
 ==========
 
+Version 1.1.0-beta05 *(2020-09-11)*
+------------------------------------
+New: Synchronizer can now be started with just a viewing key.
+New: Initializer improvments.
+New: Added tool for loading checkpoints.
+New: Added tool for deriving keys and addresses, statically.
+New: Updated and revamped the demo apps.
+New: Added a bit more (unofficial) t-addr support.
+Fix: Broken testnet demo app.
+Fix: Publishing configuration.
+
 Version 1.1.0-beta04 *(2020-08-13)*
 ------------------------------------
 New: Add support for canopy on testnet.